Completing the application, creating the feature for theme switching
Concluding the application, creating the 'temas.js' file, and using the ThemeProvider component from StyledComponents to implement the functionality of switching between light and dark themes. This involves modifying the 'variaveis.js' file by adding the style variables for each of the themes. Additionally, creating the styled components for 'IconeTema' and 'BtnTema' to modify the icons when toggling between themes, and designing a button for directly changing the themes. Finally, developing the 'SwitcherTema' component to dynamically adjust the icon tones, and utilizing the useState hook to store the state of the currently rendered theme